extract game and reusable utils.
